Remove the dependency on libncurses
Standard bash requires libncurses - it uses its tputs() function.
It includes the sources of a tiny library to provide that functionality.

This change:
- Adds a new static library - libtermcap - the tiny library mentioned above
- Removes libncurses from the list of libraries bash links with, and adds
  the new libtermcap instead

By removing this dependency, we remove the product dependency on libncurses,
and save both the concerns of its legal notices and its part in the SDK
generation.

Change-Id: I56d9f506067050807e5eda7d465f2710cf47d076